+++ andes/models/motor/motor3.py
@@ -31,7 +31,7 @@ class Motor3(MotorBaseData, Motor3Model):
     See "Power System Modelling and Scripting" by F. Milano.
 
     To simulate motor startup, set the motor status ``u`` to ``0``
-    and use a ``Toggler`` to control the model.
+    and use a ``Toggle`` to control the model.
     """
 
     def __init__(self, system, config):