+++ docs/source/release-notes.rst
@@ -12,6 +12,7 @@ v1.7 Notes
 v1.7.4 (2022-07-XX)
 -------------------
 
+- Renamed model `Toggler` to `Toggle`. 
 - New model: ESAC5A. (Contributed by Ahmad Ali).
 
 v1.7.3 (2022-06-25)